inventory component in inventorytab
