+++ src/bm-app/InventoryTab.stories.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+import Faker from "faker";
 import { InventoryTab } from "./InventoryTab";
 
 export default {
@@ -9,4 +10,28 @@ export default {
 const Template = (args) => <InventoryTab {...args} />;
 
 export const Primary = Template.bind({});
-Primary.args = {};
+Primary.args = {
+  items: Array(2)
+    .fill()
+    .map(() => ({
+      name: Faker.commerce.productName(),
+      description: Faker.commerce.productDescription(),
+    }))
+    .concat(
+      Array(2)
+        .fill()
+        .map(() => ({
+          name: Faker.commerce.productName(),
+          description: Faker.commerce.productDescription(),
+          events: Array(3)
+            .fill()
+            .map(() => ({
+              name: Faker.lorem.word(),
+              description: Faker.lorem.words(),
+              didHappen: Faker.datatype.boolean(),
+              message: Faker.lorem.sentences(),
+            })),
+        }))
+    ),
+  addEvent: () => {},
+};
